[Plays] Game needed
The Plays controller, and the routes leading to it, will be nested under
Games routes; i.e., they exist within the context of a Game. Therefore,
a Games controller, and routes leading to IT, are necessary. Also, these
nested routes will depend on a `:game_id`, and therefore a Game model
will also be necessary.

Modify the PlaysController. It should no longer refer to "sessions".

Create a Game model:

```
> rails generate model Game
      invoke  active_record
      create    db/migrate/20151122232736_create_games.rb
      create    app/models/game.rb
      invoke    rspec
      create      spec/models/game_spec.rb
      invoke      factory_girl
      create        spec/factories/games.rb
```

Rename all the Plays-related views from the "sessions" sub-directory to
the "plays" sub-directory. (also, previously, the view templates for
Games were created, but not added to the repository - do so now)

Migrate the Database after the above.

```
> rake db:migrate
```