[Plays] Starting a new Play session
Rails classes and objects can be difficult to test, or at least,
difficult to test in a way that is easy to read. A library of useful
aliases and macros, such as the Shoulda Matchers from ThoughtBot, would
go a long way to solving this problem.

To finish the "Start a play session" feature, the Game model will need
to be associated to the Play model with a one-to-many relationship. That
requirement, as well as the fact that Plays should be nested within a
Game's context, will necessitate changing the views involved to reflect
the new association.  Also, the controllers and view templates involved
will need access to a `@game` and `@play` instance variable when
rendering the requested views.

---

Add `shoulda-matchers` to the Gemfile to make testing Rails classes and
objects easier to write and read.

- e.g. `it { is_expected.to have_many :plays }`.

Also, a `support/shoulda_matchers.rb` file is created to hold the
necessary configuration, and the `rails_helper.rb` file is modified to
`require` the new support file.

---

Complete the "Start a play session" feature by modifying the included
"Dungeon Master starts a new Play session" scenario to correctly drive
the application to a new Play session's page.  To accomplish this,
modify the Game model so that it `has_many :plays`. This requires a
:game_id foreign key be added to to the "plays" table (also, migrate the
database).

```
> rails generate migration AddGameRefToPlays game:references
      invoke  active_record
      create    db/migrate/20151123001819_add_game_ref_to_plays.rb

> rake db:migrate
== 20151123001819 AddGameRefToPlays: migrating =========================
-- add_reference(:plays, :game, {:index=>true, :foreign_key=>true})
   -> 0.0311s
== 20151123001819 AddGameRefToPlays: migrated (0.0312s) ================
```

---

Misc:

* RSpec is constantly reporting on the speed of the 10 slowest tests.
  * Make this stop. (i.e. Shut up, RSpec!!)